Move home view into a seperate file
