+++ lib/main.dart
@@ -1,11 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:frontend/widgets/post.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
 
 import 'common/config.dart' as config;
 import 'common/token.dart';
-import 'widgets/timeline.dart';
-import 'widgets/users.dart';
+import 'widgets/home.dart';
 
 main() async {
   final AuthLink authLink =
@@ -30,61 +28,6 @@ class App extends StatelessWidget {
   @override
   Widget build(BuildContext context) => MaterialApp(
         title: 'FSE Frontend App',
-        home: AppHome(),
+        home: Home(),
       );
-}
-
-class AppHome extends StatefulWidget {
-  @override
-  State<StatefulWidget> createState() {
-    return _AppHomeState();
-  }
-}
-
-class _AppHomeState extends State<AppHome> {
-  int _index = 0;
-
-  final String _title = 'Full Stack Example';
-
-  final List<Widget> _views = [
-    Timeline(),
-    UserList(),
-  ];
-
-  _onTappedIcon(int index) {
-    setState(() {
-      this._index = index;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(_title),
-      ),
-      resizeToAvoidBottomPadding: false,
-      floatingActionButton: Post(),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      bottomNavigationBar: BottomAppBar(
-        shape: CircularNotchedRectangle(),
-        clipBehavior: Clip.antiAlias,
-        child: BottomNavigationBar(
-          currentIndex: _index,
-          onTap: _onTappedIcon,
-          items: [
-            BottomNavigationBarItem(
-              icon: Icon(Icons.timeline),
-              title: Text('Timeline'),
-            ),
-            BottomNavigationBarItem(
-              icon: Icon(Icons.list),
-              title: Text('Users'),
-            ),
-          ],
-        ),
-      ),
-      body: _views[_index],
-    );
-  }
 }
